FNEG feito

+++ examples/chris.java
@@ -1,6 +1,6 @@
 public class chris{
     public static void main(String[] args) {
-        long test =  15;
+        float test =  15;
         test = -test;
         System.out.println(test);
     }
